backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
python-dotenv>=1.0,<2.0
psycopg[binary]>=3.1,<4.0
